Add named actors and queue messages for non-existent actors.
The point of this is that when an actor crashes and someone is trying to
send a message to one, the messages don't immediately disappear. The scheme
used doesn't guarantee messages will eventually be delivered but makes
message loss much less likely to happen. I hope.

And of course, named actors are a convenience.