Add naming system for actors and an external message queue.
Actors can now have names. Messages can be send on name-basis. The
external message queue catches messages that were sent to non-existent
actors by name and are sent to those actors when they come to exist.

Hopefully, this lessens the chance of lost messages when messages are sent
to named actors that don't exist.

Includes a silly and simple test that it works at least to some extent.